🐛fix: db connection pool 누수 문제 해결
