+++ build.gradle
@@ -51,7 +51,6 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-data-redis' // redis
     implementation platform("io.awspring.cloud:spring-cloud-aws-dependencies:3.1.0")
     implementation 'io.awspring.cloud:spring-cloud-aws-starter-parameter-store' // AWS Parameter Store
-    implementation 'software.amazon.awssdk:sts:2.21.12'
     annotationProcessor "com.querydsl:querydsl-apt:5.0.0:jakarta"
     annotationProcessor "jakarta.annotation:jakarta.annotation-api"
     annotationProcessor "jakarta.persistence:jakarta.persistence-api"
